删除一个文件，添加一个

+++ src/main/java/com/duan/Demo1.java
@@ -1,5 +1,5 @@
 package com.duan;
 
 public class Demo1 {
-    double a = 0;
+    int a = 0;
 }
